refactor(ethereum_clis): move concrete implementations to sub-package
