+++ src/ethereum_clis/__init__.py
@@ -2,13 +2,13 @@
 Library of Python wrappers for the different implementations of transition tools.
 """
 
-from .besu import BesuTransitionTool
 from .clis import CLINotFoundInPath, UnknownCLI
-from .ethereumjs import EthereumJSTransitionTool
-from .evmone import EvmOneTransitionTool
-from .execution_specs import ExecutionSpecsTransitionTool
-from .geth import GethTransitionTool
-from .nimbus import NimbusTransitionTool
+from .clis.besu import BesuTransitionTool
+from .clis.ethereumjs import EthereumJSTransitionTool
+from .clis.evmone import EvmOneTransitionTool
+from .clis.execution_specs import ExecutionSpecsTransitionTool
+from .clis.geth import GethTransitionTool
+from .clis.nimbus import NimbusTransitionTool
 from .transition_tool import TransitionTool
 from .types import Result, TransitionToolOutput
