+++ src/ethereum_clis/clis/execution_specs.py
@@ -13,7 +13,7 @@ from typing import Optional
 
 from ethereum_test_forks import Fork
 
-from .transition_tool import TransitionTool
+from ..transition_tool import TransitionTool
 
 DAEMON_STARTUP_TIMEOUT_SECONDS = 5
